Refactored voting_results.tsx
