+++ packages/commonwealth/client/scripts/views/components/proposals/YesNoRejectResult.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import NearSputnikProposal from '../../../controllers/chain/near/sputnik/proposal';
+import { YesNoRejectVotingResult } from './voting_result_components';
+
+export function YesNoRejectResult({ votes, proposal }) {
+  return (
+    <YesNoRejectVotingResult
+      proposal={proposal as NearSputnikProposal}
+      votes={votes}
+    />
+  );
+}
