+++ packages/commonwealth/client/scripts/views/components/proposals/votingResults/DefaultVotingResult.tsx
@@ -1,0 +1,11 @@
+import { VotingResult } from '../voting_result_components';
+
+export function DefaultVotingResult({ votes, proposal }) {
+  return (
+    <VotingResult
+      yesVotes={votes.filter((v) => v.choice === true)}
+      noVotes={votes.filter((v) => v.choice === false)}
+      proposal={proposal}
+    />
+  );
+}
