+++ packages/commonwealth/client/scripts/views/pages/view_proposal/index.tsx
@@ -21,7 +21,7 @@ import type { AnyProposal } from '../../../models/types';
 import { CollapsibleProposalBody } from '../../components/collapsible_body_text';
 import { CWContentPage } from '../../components/component_kit/cw_content_page';
 import { VotingActions } from '../../components/proposals/voting_actions';
-import { VotingResults } from '../../components/proposals/voting_results';
+import { VotingResults } from '../../components/proposals/VotingResults';
 import { TipDetail } from '../tip_detail';
 import { AaveViewProposalDetail } from './aave_summary';
 import type { LinkedSubstrateProposal } from './linked_proposals_embed';
@@ -102,7 +102,6 @@ export function useProposalData(proposalId, typeProp) {
     }
   }, [isAdapterLoaded, proposalId]);
 
-  console.log(metadata);
   return {
     error,
     metadata,
